slightly diffferent functions
So this has an encryption/de-encryption algorithm that goes slightly further in that it uses an iterative mod function to encrypt a message. I'm extremely proud of my accomplishment. Figuring out how to reverse the mod function was easy as a solution, but I had to learn why my algorithm was failing me.

The main thing was that I learned that ascii uses the first 31 values for command prompts and stuff, so my algorithm would get tripped up on that. So then I got clever, eventually figuring out a mod + b function that would only go through the character values neccesary. I hope you think this is cool